+++ contracts/treasury/CHANGELOG.md
@@ -5,6 +5,26 @@ Edit the package's CHANGELOG.json file only.
 
 CHANGELOG
 
+## v1.4.53 - _August 29, 2023_
+
+    * Dependencies updated
+
+## v1.4.52 - _August 18, 2023_
+
+    * Dependencies updated
+
+## v1.4.51 - _August 9, 2023_
+
+    * Dependencies updated
+
+## v1.4.50 - _July 21, 2023_
+
+    * Dependencies updated
+
+## v1.4.49 - _July 19, 2023_
+
+    * Dependencies updated
+
 ## v1.4.48 - _May 10, 2023_
 
     * Dependencies updated